requirements-built-web.txt
FormEncode==1.3.0
Mako==1.0.1
MarkupSafe==0.23
PasteDeploy==1.5.2
Pillow==2.9.0
Pygments==2.0.2
SQLAlchemy==1.0.5
SQLAlchemy-Utils==0.30.10
Unidecode==0.04.18
WebHelpers==1.3
WebHelpers2==2.0
WebOb==1.4.1
WebTest==2.0.18
alembic==0.7.6
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
arrow==0.5.4
awesome-slugify==1.6.5
beautifulsoup4==4.3.2
billiard==3.3.0.20
celery==3.1.18
cffi==1.2.1
cryptography==1.0
decorator==3.4.2
enum34==1.0.4
futures==3.0.3
gevent==1.0.2
greenlet==0.4.7
gunicorn==19.3.0
httplib2==0.9.1
idna==2.0
ipaddress==1.0.14
kombu==3.0.26
natsort==4.0.1
ndg-httpsclient==0.4.0
nose==1.3.7
nose-selecttests==0.4
psycopg2==2.6.1
py-bcrypt==0.4
pyOpenSSL==0.15.1
pyasn1==0.1.8
pycparser==2.14
pymongo==3.0.2
pyramid==1.5.7
pyramid-celery==2.0.0
pyramid-debugtoolbar==2.4
pyramid-mako==1.0.2
pyramid-simpleform==0.6.1
pyramid-storage==0.0.8
pyramid-tm==0.12
python-dateutil==2.4.2
python-thumbnails==0.5.1
pytz==2015.4
redis==2.10.3
regex==2015.06.15
repoze.lru==0.6
requests==2.7.0
requests-futures==0.9.5
six==1.9.0
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.9
wsgiref==0.1.2
zope.deprecation==4.1.2
zope.interface==4.1.2
zope.sqlalchemy==0.7.6
